refactor(portfolio): split DAS fixture into items/ and responses/ modules
The single fixture file had grown past the readable threshold; splitting by
purpose (mainstream items, edge cases, native balance, populated/empty
responses) makes a fixture easier to find by name. The original path stays
as a thin barrel so existing imports work unchanged.
