+++ src/features/portfolio/__fixtures__/items/native.ts
@@ -1,0 +1,9 @@
+/**
+ * Native balance fixture. Shape verified from live API; values are
+ * representative test data.
+ */
+export const nativeBalanceFixture = {
+  lamports: 2_500_000_000n,
+  price_per_sol: 82.00442504882812,
+  total_price: 205.01106262207031,
+}
