+++ src/features/portfolio/__fixtures__/items/owner.ts
@@ -1,0 +1,2 @@
+/** Owner wallet address used across the fixture. Synthetic. */
+export const FIXTURE_OWNER = '7EcDhSYGxXyscszYEp35KHN8vvw3svAuLKTzXwCFLtV'
